requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
mysql-connector-python==8.0.31
requests==2.27.1
